services/hass_input_helper/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
jinja2==3.1.3
paho-mqtt==1.6.1
